requirements.txt
appdirs==1.4.3
asgiref==3.2.3
boto3==1.12.48
botocore==1.15.48
certifi==2020.4.5.1
chardet==3.0.4
distlib==0.3.0
dj-database-url==0.5.0
Django==2.0.2
django-heroku==0.3.1
django-payu==0.5
django-storages==1.9.1
docutils==0.15.2
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
jmespath==0.9.5
Pillow==6.2.1
pipenv==2018.11.26
psycopg2==2.8.4
pycryptodome==3.9.7
python-dateutil==2.8.1
pytz==2019.3
razorpay==1.2.0
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
virtualenv==20.0.18
virtualenv-clone==0.5.4
whitenoise==5.0.1
zipp==3.1.0